feat: Taking one step back and thinking of something simple.
