+++ go.mod
@@ -4,5 +4,5 @@ go 1.16
 
 require (
 	go.uber.org/multierr v1.6.0 // indirect
-	go.uber.org/zap v1.16.0 // indirect
+	go.uber.org/zap v1.16.0
 )
