+++ internal/listing/package.go
@@ -1,0 +1,3 @@
+package listing
+
+type Package string
