+++ internal/listing/service.go
@@ -1,0 +1,6 @@
+package listing
+
+type Repository interface {
+	// GetAllPackages returns all packages saved in storage.
+	GetAllPackages() []Package
+}
