--- pkg/listing/package.go
@@ -1,9 +1,0 @@
-package listing
-
-type Distribution string
-
-type Package struct {
-	Name          string
-	Version       string
-	Distributions []Distribution
-}
