--- pkg/storage/disk/package.go
@@ -1,6 +1,0 @@
-package disk
-
-type Package struct {
-	Name    string
-	Version string
-}
